Commented out a check that I'm not sure is needed for starting out.
The check is for the length of pmid array that will be added to db. It assumes
that if the pmid length is the same as the current size of the references field then these pmids can't be new refs. I think this came in handy after having collected citations so will revisit once document citation collection.