articles.endpoint.unit.tests passes but with db connection error
Connection error is due to mongoose.connect which is for one database.
Somewhere two connections are being made which to handle mongoose.createConnection should be used.
However using createConnection requires a refactor where the models will be created using ArticleModel db.model('ArticleModel...) where db is an instance of the database. Currently ArticleModel is connecting directly to the database.

Ultimately this may require being able to pass dev, test and production environments to the app.