Test for #getQueue now passing.
This test was going to initQueue and return 130,000 objects from the database. This took a while to say the least so updated the query to findOne instead.

Means I can go back later and probably delete initQueue and getQueue and just have getPmid. Abandoning the idea of getting a queue once from the database because a) holding in memory an object with 130,000 objects is not necessary and b) is too slow when database size increases.